Follow redirects to download patch URLs
For example github releases now don't directly link the files anymore,
but instead redirect you first to a different url where the release
asset is. This can be annoying and I'd prefer it to follow redirects
when downloading patch files. The curl `--location` option does this and
allows downloads from GH Releases to still work directly from the linked
URL on the release page.